add basic config in CartSideBar
